nexus-core: Mock out NetworkInterface in SystemInformationGeneratorImplTest
NetworkInterface.getInterfaces() was being called directly which will
return the interfaces available on the machine running the test. The
test assumes that all interfaces will have at least one address but that
is not always the case.  Since the test is focused on exercising
SystemInformationGeneratorImpl.reportNetworkInterface(), create a mock
NetworkInterface so it returns consistent results.