remove ref to issues.sonatype.org from issue-template.md
issues.sonatype.org is no more:

On Jan 9th, 2024 we announced issues.sonatype.org will be decommissioned soon. As part of the decommissioning process we removed the access to issues.sonatype.org, and we replaced our Jira ticketing system with a Zendesk email intake method. To open a support ticket with us email Central Support and follow-up in the email thread. To help us get back to you faster, please do not include previous replies in your follow-up messages.