Fix R repository browsing by comparing format with string equals method
Ognl does not normally compare strings with an "==" operator, only numbers can be compared this way (as in Java). Instead a comparison method like "equals" has to be used. This may have worked accidentally when both sides are clearly not numbers. Unfortunately the string "r" (for an R repository) is short enough to possible count as a character in this context, which is usable with "==", while 'conan' is clearly no such thing. As a result R repositories may not be browsable.